requirements.txt
psutil==6.1.0
prometheus_client==0.21.1
python-dotenv==1.0.1
